nativesdk-qt4-tools: ensure --disable-static is not passed
This recipe presumes it can build the qt4-tools as static, and does not
install any of it's built libraries. If we don't remove
`--disable-static`, it attempts to use non-existent shared libraries
(many of them aren't even built), resulting in either link failures at
build time or (occasionally) at populate_sdk time due to the lack of a
provider for the shared libraries.

Signed-off-by: Cody P Schafer <dev@codyps.com>