qt4-embedded: Default to build tslib when touchscreen is defined
When "touchscreen" is in the MACHINE_FEATURES, packagegroup-core-qt4e
will RDEPEND on qt4-embedded-plugin-mousedriver-tslib, but that
library is not being built by default, resulting in an error like:
opkg_prepare_url_for_install: Couldn't find anything to satisfy 'qt4-embedded-plugin-mousedriver-tslib'.

To prevent that from happening, add "tslib" to the default
PACKAGECONFIG when touchscreen is defined.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>