qt4: disable openssl support
Qt4 does not support openssl 1.1 (and likely never will), while its
oe-core dependencies are being moved to that version, which is causing a
sysroot conflict between openssl 1.0 and 1.1 headers. The only way to resolve
it is to disable openssl support in Qt4, as we don't want Qt4's outdated version
requirements for openssl to be imposed on the rest of oe-core.

Signed-off-by: Alexander Kanavin <alexander.kanavin@linux.intel.com>